requirements.txt
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
python-telegram-bot==13.14